public-api/requirements.txt
flask==2.0.2
mysql-connector-python==8.0.27